Fix circleci by moving infra tests
